cli/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
check-wheel-contents==0.3.4
click==8.0.3
docopt==0.6.2
-e git+https://github.com/evasvarna/helloworld-cli.git@e1f82c084bad160b8568ae23bcc04a46bffe0bb2#egg=helloworld_cli
idna==3.3
mysql-connector-python==8.0.28
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pipreqs==0.4.11
protobuf==3.19.4
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
tomli==2.0.1
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wheel-filename==1.4.0
yarg==0.1.9
